feat: Added confirm email endpoint
